Updated the vertex attributes for the model, now the attributes are smaller in size

+++ crates/rendering/src/utils/assets_impls.rs
@@ -57,17 +57,17 @@ impl Asset for Model {
                 }
                 // Normals
                 "n" => {
-                    let coords: Vec<f32> = other.split('/').map(|coord| coord.parse::<f32>().unwrap()).collect();
+                    let coords: Vec<i8> = other.split('/').map(|coord| coord.parse::<i8>().unwrap()).collect();
                     model.normals.push(veclib::Vector3::new(coords[0], coords[1], coords[2]));
                 }
                 // UVs
                 "u" => {
-                    let coords: Vec<f32> = other.split('/').map(|coord| coord.parse::<f32>().unwrap()).collect();
+                    let coords: Vec<u8> = other.split('/').map(|coord| coord.parse::<u8>().unwrap()).collect();
                     model.uvs.push(veclib::Vector2::new(coords[0], coords[1]));
                 }
                 // Tangents
                 "t" => {
-                    let coords: Vec<f32> = other.split('/').map(|coord| coord.parse::<f32>().unwrap()).collect();
+                    let coords: Vec<i8> = other.split('/').map(|coord| coord.parse::<i8>().unwrap()).collect();
                     model.tangents.push(veclib::Vector4::new(coords[0], coords[1], coords[2], coords[3]));
                 }
                 // Triangle indices
